@@ -3,7 +3,9 @@ description: A new Flutter package project.
 version: 0.0.1
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: ^3.5.3
+  flutter: ">=1.17.0"
+
 
 dependencies:
   intl: ^0.19.0
@@ -18,6 +20,8 @@ dependency_overrides:
   web: ^1.0.0
   json_annotation: ^4.9.0
   sqlite3: ^2.7.2
+  device_info_plus: ^11.2.2
+  
   
 dev_dependencies:
   flutter_test:
