@@ -22,7 +22,7 @@ dependencies:
   flex_color_scheme: ^4.2.0
   # image_picker: ^0.8.7+5
   file_picker: ^8.3.1
-  firebase_ui_auth: ^1.16.0
+  firebase_ui_auth: ^1.16.1
   stacked: ^3.4.2
   azlistview: ^2.0.0
   # I will abandon qr_code_scanner and switch to mobile_scanner even though this
@@ -40,7 +40,7 @@ dependencies:
   url_launcher: ^6.3.0
   google_fonts: ^6.1.0
   number_display: ^3.0.0
-  firebase_core: ^3.10.0
+  firebase_core: ^3.11.0
   cupertino_icons: ^1.0.4
   flipper_services: any
 
@@ -79,7 +79,7 @@ dependencies:
   badges: ^3.1.2
   community_charts_flutter: ^1.0.2
   
-  dropdown_search: ^5.0.6
+  dropdown_search: ^6.0.2
   loading_animation_widget: ^1.2.0+4
   email_validator_flutter: ^1.0.0
   #TODO: because this is community package, I have applied for community licence
@@ -101,9 +101,9 @@ dependencies:
   sizer: ^2.0.15
   talker_flutter: ^4.1.5
   collection: ^1.18.0
-  amplify_api: ^2.5.0
-  amplify_flutter: ^2.5.0
-  amplify_auth_cognito: ^2.5.0
+  amplify_api: ^2.6.0
+  amplify_flutter: ^2.6.0
+  amplify_auth_cognito: ^2.6.0
   flutter_side_menu: ^0.5.3
   
   phone_numbers_parser: ^8.3.0
@@ -124,6 +124,8 @@ dependency_overrides:
   web: ^1.0.0
   json_annotation: ^4.9.0
   sqlite3: ^2.7.2
+  device_info_plus: ^11.2.2
+  
 
 flutter:
   # The following line ensures that the Material Icons font is
