@@ -10,9 +10,9 @@ environment:
   sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
-  firebase_ui_auth: ^1.16.0
+  firebase_ui_auth: ^1.16.1
   flutter_form_bloc: 0.31.0-dev
-  firebase_core: ^3.10.0
+  firebase_core: ^3.11.0
   flipper_services: any
   flipper_dashboard: any
   universal_platform: ^1.0.0+1
@@ -21,11 +21,11 @@ dependencies:
   flutter_settings_ui: ^3.0.1
   stacked: ^3.4.2
   google_fonts: ^6.1.0
-  firebase_auth: ^5.4.0
+  firebase_auth: ^5.4.2
   flipper_routing: ^1.0.0
   connectivity_plus: ^6.0.5
-  firebase_ui_localizations: ^1.9.0
-  firebase_ui_oauth: ^1.5.3
+  firebase_ui_localizations: ^1.13.1
+  firebase_ui_oauth: ^1.6.1
   firebase_ui_oauth_google: ^1.4.0
   flutter_svg: ^2.0.0
   gap: ^2.0.0
@@ -38,11 +38,11 @@ dependencies:
   riverpod: ^2.0.0
   hooks_riverpod: ^2.0.0
   loading_animation_widget: ^1.2.0+4
-  sentry: ^8.6.0
-  sentry_flutter: ^8.10.1
+  sentry: ^8.13.0
+  sentry_flutter: ^8.13.0
   flipper_models: any
   # realm: ^20.0.0
-  google_sign_in: ^6.2.1
+  google_sign_in: ^6.2.2
   firebase_ui_oauth_apple: ^1.3.0
 
   
@@ -52,6 +52,8 @@ dependency_overrides:
   web: ^1.0.0
   json_annotation: ^4.9.0
   sqlite3: ^2.7.2
+  device_info_plus: ^11.2.2
+  
   
   
 flutter:
