@@ -12,7 +12,7 @@ dependencies:
   meta: ^1.3.0
   universal_platform: ^1.0.0+1
   stacked_services: ^1.6.0
-  firebase_auth: ^5.4.0
+  firebase_auth: ^5.4.2
   stacked: ^3.4.2
   built_value: ^8.1.4
   azlistview: ^2.0.0
@@ -22,8 +22,8 @@ dependencies:
   flipper_routing: ^1.0.0
   connectivity_plus: ^6.0.5
   
-  sentry: ^8.6.0
-  sentry_flutter: ^8.10.1
+  sentry: ^8.13.0
+  sentry_flutter: ^8.13.0
   
   path_provider: ^2.1.5
   http: ^1.2.2
@@ -34,10 +34,7 @@ dependencies:
   riverpod: ^2.0.0
   hooks_riverpod: ^2.0.0
   freezed_annotation: ^2.4.4
-  # cloud_firestore: ^5.4.3
   isolated_worker: ^0.1.1
-  # adding realm to replace firestore on desktop in transporting data to cloud
-  # realm: ^20.0.0
   dart_mappable: ^4.2.1
   dio: ^5.4.3+1
   talker_dio_logger: ^4.1.5
@@ -46,14 +43,14 @@ dependencies:
   rxdart: ^0.28.0
   collection: ^1.19.0
   intl: ^0.19.0
-  amplify_core: ^2.5.0
-  amplify_api: ^2.5.0
-  amplify_flutter: ^2.5.0
-  amplify_auth_cognito: ^2.5.0
-  amplify_storage_s3: ^2.5.0
+  amplify_core: ^2.6.0
+  amplify_api: ^2.6.0
+  amplify_flutter: ^2.6.0
+  amplify_auth_cognito: ^2.6.0
+  amplify_storage_s3: ^2.6.0
   async: ^2.11.0
   mime: ^1.0.6
-  firebase_core: ^3.10.0
+  firebase_core: ^3.11.0
   sqlite3_flutter_libs: ^0.5.29
   permission_handler: ^11.3.1
   cloud_firestore_odm: ^1.0.0-dev.88
@@ -64,10 +61,11 @@ dependencies:
   # cbl_flutter: ^3.1.3
   supabase_models: any
   
-  printing: ^5.13.4
+  printing: ^5.14.0
   pdf: ^3.11.1
   injectfy: ^1.0.1
   mock_supabase_http_client: 0.0.3+1
+  sqlite3_web: ^0.3.0
   
 
 dependency_overrides:
@@ -78,6 +76,8 @@ dependency_overrides:
   json_annotation: ^4.9.0
   sqlite3: ^2.7.2
   stacked_services: ^1.6.0
+  device_info_plus: ^11.2.2
+  
   # amplify_db_common:
   #   git:
   #     url: https://github.com/aws-amplify/amplify-flutter.git
@@ -89,8 +89,6 @@ dependency_overrides:
 dev_dependencies:
   build_runner: ^2.4.14
   dart_mappable_builder: ^4.2.1
-  # custom_lint: ^0.6.4
-  mockito: ^5.4.4
   json_serializable: ^6.8.0 
   cloud_firestore_odm_generator: ^1.0.0-dev.89
   riverpod_generator: ^2.6.3
