@@ -1,3 +1,4 @@
+import 'package:flipper_models/providers/variants_provider.dart';
 import 'package:flipper_services/proxy.dart';
 import 'package:flipper_ui/flipper_ui.dart';
 import 'package:flutter/material.dart';
@@ -5,6 +6,10 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:supabase_models/brick/models/all_models.dart';
 import 'package:syncfusion_flutter_datagrid/datagrid.dart';
 import 'package:talker_flutter/talker_flutter.dart';
+import 'package:dropdown_search/dropdown_search.dart';
+
+final selectedVariantProvider =
+    StateProvider.family<Variant?, String>((ref, variantId) => null);
 
 class DataRowWidget extends StatefulHookConsumerWidget {
   const DataRowWidget({
@@ -21,25 +26,23 @@ class DataRowWidget extends StatefulHookConsumerWidget {
   final TextEditingController nameController;
   final TextEditingController supplyPriceController;
   final TextEditingController retailPriceController;
-  final void Function(Variant? selectedItem) selectSale;
+  final void Function(
+    Variant? itemToAssign,
+    Variant? itemFromPurchase,
+  ) selectSale;
   final List<Variant> finalSalesList;
   final VoidCallback saveItemName;
   final void Function(List<Variant> acceptedVariants, String pchsSttsCd)
       acceptPurchases;
 
   @override
-  _DataRowWidgetState createState() => _DataRowWidgetState();
+  ConsumerState<DataRowWidget> createState() => _DataRowWidgetState();
 }
 
 class _DataRowWidgetState extends ConsumerState<DataRowWidget> {
-  // Maps to store edited retail and supply prices
   final Map<String, double> _editedRetailPrices = {};
   final Map<String, double> _editedSupplyPrices = {};
-
-  // Talker instance for logging
   final Talker talker = TalkerFlutter.init();
-
-  // Data source for the SfDataGrid
   late _DataSource _dataSource;
 
   @override
@@ -63,15 +66,18 @@ class _DataRowWidgetState extends ConsumerState<DataRowWidget> {
 
   @override
   Widget build(BuildContext context) {
+    final variantProviders =
+        ref.watch(variantProvider(branchId: ProxyService.box.getBranchId()!));
+
     return Card(
       elevation: 0,
       shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
       child: Container(
         width: double.infinity,
         child: widget.finalSalesList.isEmpty
-            ? Center(
+            ? const Center(
                 child: Padding(
-                  padding: const EdgeInsets.all(16.0),
+                  padding: EdgeInsets.all(16.0),
                   child: Text(
                     'No Data Available',
                     style: TextStyle(
@@ -93,7 +99,7 @@ class _DataRowWidgetState extends ConsumerState<DataRowWidget> {
                   if (details.rowColumnIndex.rowIndex > 0) {
                     final item = widget
                         .finalSalesList[details.rowColumnIndex.rowIndex - 1];
-                    _showEditDialog(context, item);
+                    _showEditDialog(context, item, variants: variantProviders);
                   }
                 },
               ),
@@ -101,20 +107,21 @@ class _DataRowWidgetState extends ConsumerState<DataRowWidget> {
     );
   }
 
-  void _showEditDialog(BuildContext context, Variant item) {
+  void _showEditDialog(BuildContext context, Variant item,
+      {required AsyncValue<List<Variant>> variants}) {
     showDialog(
       context: context,
       builder: (context) => Dialog(
         shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
         child: SingleChildScrollView(
           child: Container(
-            constraints: const BoxConstraints(maxWidth: 600),
+            constraints: const BoxConstraints(maxWidth: 890),
             padding: const EdgeInsets.all(16),
             child: Column(
               mainAxisSize: MainAxisSize.min,
               children: [
                 const SizedBox(height: 16),
-                _buildItemsTable(item),
+                _buildItemsTable(item, variants: variants),
                 const SizedBox(height: 24),
                 _buildDialogActions(context),
               ],
@@ -123,35 +130,33 @@ class _DataRowWidgetState extends ConsumerState<DataRowWidget> {
         ),
       ),
     ).then((_) {
-      // Update the DataGrid when the dialog is closed
       _updateDataGrid();
     });
   }
 
-  Widget _buildItemsTable(Variant item) {
+  Widget _buildItemsTable(Variant item,
+      {required AsyncValue<List<Variant>> variants}) {
     return Card(
       elevation: 2,
       shape: RoundedRectangleBorder(
         borderRadius: BorderRadius.circular(16),
       ),
-      child: Container(
-        width: double.infinity,
+      child: Padding(
         padding: const EdgeInsets.all(20),
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            // Table Header Section
-            Text(
-              'Item Details',
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.bold,
-                color: Colors.grey.shade800,
+            Padding(
+              padding: const EdgeInsets.only(bottom: 20),
+              child: Text(
+                'Item Details',
+                style: TextStyle(
+                  fontSize: 20,
+                  fontWeight: FontWeight.bold,
+                  color: Colors.grey.shade800,
+                ),
               ),
             ),
-            const SizedBox(height: 16),
-
-            // Data Table
             Theme(
               data: Theme.of(context).copyWith(
                 dividerColor: Colors.grey.shade200,
@@ -165,115 +170,196 @@ class _DataRowWidgetState extends ConsumerState<DataRowWidget> {
                     color: Colors.black87,
                     fontSize: 14,
                   ),
+                  headingRowColor:
+                      WidgetStateProperty.all(Colors.grey.shade100),
+                  horizontalMargin: 16,
                 ),
               ),
-              child: DataTable(
-                headingRowHeight: 56,
-                horizontalMargin: 12,
-                columnSpacing: 16,
-                headingRowColor: WidgetStateProperty.all(Colors.grey.shade100),
+              child: Container(
                 decoration: BoxDecoration(
                   border: Border.all(color: Colors.grey.shade200),
                   borderRadius: BorderRadius.circular(12),
                 ),
-                columns: const [
-                  DataColumn(
-                    label: Text('Item Name'),
-                    tooltip: 'Item Name',
-                  ),
-                  DataColumn(
-                    label: Text('Qty'),
-                    tooltip: 'Quantity',
-                  ),
-                  DataColumn(
-                    label: Text('Supply'),
-                    tooltip: 'Supply Price',
-                  ),
-                  DataColumn(
-                    label: Text('Retail'),
-                    tooltip: 'Retail Price',
-                  ),
-                ],
-                rows: [
-                  DataRow(
-                    cells: [
-                      DataCell(
-                        Row(
-                          mainAxisSize: MainAxisSize.min,
-                          children: [
-                            Container(
-                              width: 32,
-                              height: 32,
-                              decoration: BoxDecoration(
-                                color: Colors.blue.shade50,
-                                borderRadius: BorderRadius.circular(6),
-                              ),
-                              child: Icon(
-                                Icons.inventory_2,
-                                color: Colors.blue.shade400,
-                                size: 18,
-                              ),
-                            ),
-                            const SizedBox(width: 8),
-                            Flexible(
-                              child: Text(
-                                item.name,
-                                style: const TextStyle(
-                                    fontWeight: FontWeight.w500),
-                                overflow: TextOverflow.ellipsis,
-                              ),
-                            ),
-                          ],
+                child: DataTable(
+                  headingRowHeight: 56,
+                  columnSpacing: 24,
+                  columns: const [
+                    DataColumn(
+                      label: Expanded(
+                        child: Text(
+                          'Item Name',
+                          style: TextStyle(fontWeight: FontWeight.bold),
                         ),
                       ),
-                      DataCell(
-                        Container(
-                          padding: const EdgeInsets.symmetric(
-                              horizontal: 8, vertical: 4),
-                          decoration: BoxDecoration(
-                            color: Colors.green.shade50,
-                            borderRadius: BorderRadius.circular(16),
-                          ),
-                          child: Text(
-                            item.stock!.currentStock.toString(),
-                            style: TextStyle(
-                              color: Colors.green.shade700,
-                              fontWeight: FontWeight.w500,
+                    ),
+                    DataColumn(
+                      label: Expanded(
+                        child: Text(
+                          'Qty',
+                          style: TextStyle(fontWeight: FontWeight.bold),
+                        ),
+                      ),
+                    ),
+                    DataColumn(
+                      label: Expanded(
+                        child: Text(
+                          'Supply',
+                          style: TextStyle(fontWeight: FontWeight.bold),
+                        ),
+                      ),
+                    ),
+                    DataColumn(
+                      label: Expanded(
+                        child: Text(
+                          'Retail',
+                          style: TextStyle(fontWeight: FontWeight.bold),
+                        ),
+                      ),
+                    ),
+                    DataColumn(
+                      label: Expanded(
+                        child: Text(
+                          'Item to assign',
+                          style: TextStyle(fontWeight: FontWeight.bold),
+                        ),
+                      ),
+                    ),
+                  ],
+                  rows: [
+                    DataRow(
+                      cells: [
+                        DataCell(
+                          Padding(
+                            padding: const EdgeInsets.symmetric(vertical: 8),
+                            child: Row(
+                              mainAxisSize: MainAxisSize.min,
+                              children: [
+                                Container(
+                                  width: 36,
+                                  height: 36,
+                                  decoration: BoxDecoration(
+                                    color: Colors.blue.shade50,
+                                    borderRadius: BorderRadius.circular(8),
+                                  ),
+                                  child: Center(
+                                    child: Icon(
+                                      Icons.inventory_2,
+                                      color: Colors.blue.shade400,
+                                      size: 20,
+                                    ),
+                                  ),
+                                ),
+                                const SizedBox(width: 12),
+                                Flexible(
+                                  child: Text(
+                                    item.name,
+                                    style: const TextStyle(
+                                      fontWeight: FontWeight.w500,
+                                      fontSize: 14,
+                                    ),
+                                    overflow: TextOverflow.ellipsis,
+                                  ),
+                                ),
+                              ],
                             ),
                           ),
                         ),
-                      ),
-                      DataCell(_buildPriceTextField(
-                        initialValue:
-                            _editedSupplyPrices[item.id]?.toString() ??
-                                item.supplyPrice?.toString() ??
-                                '',
-                        hintText: 'Supply',
-                        onChanged: (value) {
-                          final supplyPrice = double.tryParse(value);
-                          if (supplyPrice != null) {
-                            _editedSupplyPrices[item.id] = supplyPrice;
-                            _updateDataGrid();
-                          }
-                        },
-                      )),
-                      DataCell(_buildPriceTextField(
-                        initialValue:
-                            _editedRetailPrices[item.id]?.toString() ??
-                                item.retailPrice?.toString() ??
-                                '',
-                        hintText: 'Retail',
-                        onChanged: (value) {
-                          final retailPrice = double.tryParse(value);
-                          if (retailPrice != null) {
-                            _editedRetailPrices[item.id] = retailPrice;
-                            _updateDataGrid();
-                          }
-                        },
-                      )),
-                    ],
-                  ),
-                ],
+                        DataCell(
+                          Container(
+                            padding: const EdgeInsets.symmetric(
+                              horizontal: 12,
+                              vertical: 6,
+                            ),
+                            decoration: BoxDecoration(
+                              color: Colors.green.shade50,
+                              borderRadius: BorderRadius.circular(20),
+                            ),
+                            child: Text(
+                              item.stock!.currentStock.toString(),
+                              style: TextStyle(
+                                color: Colors.green.shade700,
+                                fontWeight: FontWeight.w500,
+                                fontSize: 14,
+                              ),
+                            ),
+                          ),
+                        ),
+                        DataCell(
+                          Padding(
+                            padding: const EdgeInsets.symmetric(vertical: 4),
+                            child: _buildPriceTextField(
+                              initialValue:
+                                  _editedSupplyPrices[item.id]?.toString() ??
+                                      item.supplyPrice?.toString() ??
+                                      '',
+                              hintText: 'Supply',
+                              onChanged: (value) {
+                                final supplyPrice = double.tryParse(value);
+                                if (supplyPrice != null) {
+                                  _editedSupplyPrices[item.id] = supplyPrice;
+                                  _updateDataGrid();
+                                }
+                              },
+                            ),
+                          ),
+                        ),
+                        DataCell(
+                          Padding(
+                            padding: const EdgeInsets.symmetric(vertical: 4),
+                            child: _buildPriceTextField(
+                              initialValue:
+                                  _editedRetailPrices[item.id]?.toString() ??
+                                      item.retailPrice?.toString() ??
+                                      '',
+                              hintText: 'Retail',
+                              onChanged: (value) {
+                                final retailPrice = double.tryParse(value);
+                                if (retailPrice != null) {
+                                  _editedRetailPrices[item.id] = retailPrice;
+                                  _updateDataGrid();
+                                }
+                              },
+                            ),
+                          ),
+                        ),
+                        DataCell(
+                          Consumer(builder: (context, ref, _) {
+                            final selectedVariant =
+                                ref.watch(selectedVariantProvider(item.id));
+                            final setSelectedVariant = ref.read(
+                                selectedVariantProvider(item.id).notifier);
+
+                            return variants.when(
+                              data: (variants) {
+                                return DropdownSearch<Variant>(
+                                  selectedItem: selectedVariant,
+                                  // Provide the list of items directly.
+                                  items: (a, b) => variants,
+                                  // Add compareFn to compare Variant objects correctly
+                                  compareFn: (Variant i, Variant s) =>
+                                      i.id == s.id,
+                                  itemAsString: (Variant v) => v.name,
+                                  decoratorProps: const DropDownDecoratorProps(
+                                    baseStyle: TextStyle(fontSize: 13),
+                                    decoration: InputDecoration(
+                                      border: OutlineInputBorder(),
+                                    ),
+                                  ),
+                                  onChanged: (Variant? itemToAssign) {
+                                    setSelectedVariant.state = itemToAssign;
+                                    widget.selectSale(itemToAssign, item);
+                                  },
+                                );
+                              },
+                              loading: () => const Text("Loading variants..."),
+                              error: (error, stack) => Text('Error: $error'),
+                            );
+                          }),
+                        ),
+                      ],
+                    ),
+                  ],
+                ),
               ),
             ),
           ],
@@ -358,8 +444,6 @@ class _DataRowWidgetState extends ConsumerState<DataRowWidget> {
               variant.supplyPrice = supplyPrice;
               ProxyService.strategy.updateVariant(updatables: [variant]);
             });
-
-            //widget.saveItemName();
 
             talker.log(
                 'Prices updated for items: ${_editedRetailPrices.keys.join(', ')}');
@@ -521,19 +605,19 @@ class _DataSource extends DataGridSource {
 
     switch (variant.pchsSttsCd) {
       case "01":
-        badgeColor = Colors.orange.withOpacity(0.2);
+        badgeColor = Colors.orange.withValues(alpha: .2);
         statusText = 'Pending';
         break;
       case "02":
-        badgeColor = Colors.green.withOpacity(0.2);
+        badgeColor = Colors.green.withValues(alpha: 0.2);
         statusText = 'Accepted';
         break;
       case "04":
-        badgeColor = Colors.red.withOpacity(0.2);
+        badgeColor = Colors.red.withValues(alpha: 0.2);
         statusText = 'Canceled';
         break;
       default:
-        badgeColor = Colors.grey.withOpacity(0.2);
+        badgeColor = Colors.grey.withValues(alpha: 0.2);
         statusText = 'Unknown';
     }
 
@@ -546,7 +630,7 @@ class _DataSource extends DataGridSource {
       child: Text(
         statusText,
         style: TextStyle(
-          color: badgeColor.withOpacity(1.0),
+          color: badgeColor.withValues(alpha: 1.0),
           fontWeight: FontWeight.bold,
         ),
       ),
@@ -557,7 +641,6 @@ class _DataSource extends DataGridSource {
     final variant = finalSalesList.firstWhere((v) => v.id == id);
     variant.pchsSttsCd = status;
 
-    // Call the acceptPurchases callback with the updated list of variants
     acceptPurchases([variant], status);
 
     talker.log(
