@@ -1,9 +1,13 @@
 import 'package:flipper_dashboard/DataRow.dart';
+import 'package:flipper_models/providers/outer_variant_provider.dart';
+import 'package:flipper_models/providers/scan_mode_provider.dart';
+import 'package:flipper_services/proxy.dart';
 import 'package:flipper_ui/flipper_ui.dart';
 import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:supabase_models/brick/models/all_models.dart';
 
-class PurchaseSaleWidget extends StatefulWidget {
+class PurchaseSaleWidget extends StatefulHookConsumerWidget {
   final GlobalKey<FormState> formKey;
   final TextEditingController nameController;
   final TextEditingController supplyPriceController;
@@ -12,7 +16,10 @@ class PurchaseSaleWidget extends StatefulWidget {
   final void Function(
       {required List<Variant> variants,
       required String pchsSttsCd}) acceptPurchases;
-  final void Function(Variant? selectedItem) selectSale;
+  final void Function(
+    Variant? itemToAssign,
+    Variant? itemFromPurchase,
+  ) selectSale;
   final List<Variant> finalSalesList;
 
   PurchaseSaleWidget({
@@ -30,7 +37,7 @@ class PurchaseSaleWidget extends StatefulWidget {
   _PurchaseSaleWidgetState createState() => _PurchaseSaleWidgetState();
 }
 
-class _PurchaseSaleWidgetState extends State<PurchaseSaleWidget> {
+class _PurchaseSaleWidgetState extends ConsumerState<PurchaseSaleWidget> {
   @override
   Widget build(BuildContext context) {
     return Container(
@@ -65,10 +72,27 @@ class _PurchaseSaleWidgetState extends State<PurchaseSaleWidget> {
                     supplyPriceController: widget.supplyPriceController,
                     retailPriceController: widget.retailPriceController,
                     saveItemName: widget.saveItemName,
-                    acceptPurchases: (List<Variant> variants, pchsSttsCd) =>
-                        widget.acceptPurchases(
-                            variants: variants, pchsSttsCd: pchsSttsCd),
-                    selectSale: (variant) => widget.selectSale(variant),
+                    acceptPurchases:
+                        (List<Variant> variants, pchsSttsCd) async {
+                      widget.acceptPurchases(
+                          variants: variants, pchsSttsCd: pchsSttsCd);
+                      // Refresh the product list
+                      // Refresh the product list
+                      ref
+                          .read(searchStringProvider.notifier)
+                          .emitString(value: "search");
+                      ref
+                          .read(searchStringProvider.notifier)
+                          .emitString(value: "");
+
+                      await ref
+                          .read(
+                              productsProvider(ProxyService.box.getBranchId()!)
+                                  .notifier)
+                          .loadProducts(searchString: "", scanMode: true);
+                    },
+                    selectSale: (itemToAssign, itemFromPurchase) =>
+                        widget.selectSale(itemToAssign, itemFromPurchase),
                     finalSalesList: widget.finalSalesList,
                   ),
                 ],
