@@ -5,6 +5,7 @@ import 'package:flipper_dashboard/TextEditingControllersMixin.dart';
 import 'package:flipper_dashboard/TransactionItemTable.dart';
 import 'package:flipper_dashboard/payable_view.dart';
 import 'package:flipper_dashboard/mixins/previewCart.dart';
+import 'package:flipper_models/providers/pay_button_provider.dart';
 import 'package:flipper_models/realm_model_export.dart';
 import 'package:flipper_models/view_models/mixins/_transaction.dart';
 import 'package:flipper_models/view_models/mixins/riverpod_states.dart';
@@ -159,9 +160,7 @@ class _QuickSellingViewState extends ConsumerState<QuickSellingView>
                 transactionAsyncValue.whenData((ITransaction transaction) {
                   startCompleteTransactionFlow(
                       immediateCompletion: imediteCompleteTransaction,
-                      completeTransaction: () {
-                        ref.read(loadingProvider.notifier).stopLoading();
-                      },
+                      completeTransaction: () {},
                       transaction: transaction,
                       paymentMethods: ref.watch(paymentMethodsProvider));
                 });
@@ -340,14 +339,14 @@ class _QuickSellingViewState extends ConsumerState<QuickSellingView>
         }
         final number = double.tryParse(value);
         if (number == null) {
-          ref.read(loadingProvider.notifier).stopLoading();
+          ref.read(payButtonLoadingProvider.notifier).stopLoading();
           return 'Please enter a valid number';
         }
 
         /// this is a percentage not amount as this percenage will be applicable
         /// to the whole item on cart, currently we only support discount on whole total
         if (number < 0 || number > 100) {
-          ref.read(loadingProvider.notifier).stopLoading();
+          ref.read(payButtonLoadingProvider.notifier).stopLoading();
           return 'Discount must be between 0 and 100';
         }
         return null;
@@ -397,16 +396,16 @@ class _QuickSellingViewState extends ConsumerState<QuickSellingView>
       }),
       validator: (String? value) {
         if (value == null || value.isEmpty) {
-          ref.read(loadingProvider.notifier).stopLoading();
+          ref.read(payButtonLoadingProvider.notifier).stopLoading();
           return 'Please enter received amount';
         }
         final number = double.tryParse(value);
         if (number == null) {
-          ref.read(loadingProvider.notifier).stopLoading();
+          ref.read(payButtonLoadingProvider.notifier).stopLoading();
           return 'Please enter a valid number';
         }
         if (number < totalAfterDiscountAndShipping) {
-          ref.read(loadingProvider.notifier).stopLoading();
+          ref.read(payButtonLoadingProvider.notifier).stopLoading();
           return 'You are receiving less than the total due';
         }
         return null;
@@ -459,12 +458,10 @@ class _QuickSellingViewState extends ConsumerState<QuickSellingView>
           .writeString(key: 'currentSaleCustomerPhoneNumber', value: value),
       validator: (String? value) {
         if (value == null || value.isEmpty) {
-          ref.read(loadingProvider.notifier).stopLoading();
           return 'Please enter a phone number';
         }
         final phoneExp = RegExp(r'^[1-9]\d{8}$');
         if (!phoneExp.hasMatch(value)) {
-          ref.read(loadingProvider.notifier).stopLoading();
           return 'Please enter a valid 9-digit phone number without a leading zero';
         }
         return null;
