@@ -7,17 +7,15 @@
 #include "generated_plugin_registrant.h"
 
 #include <amplify_db_common/amplify_db_common_plugin.h>
-#include <cbl_flutter_ce/cbl_flutter_ce.h>
 #include <desktop_webview_auth/desktop_webview_auth_plugin.h>
 #include <file_selector_linux/file_selector_plugin.h>
 #include <flutter_localization/flutter_localization_plugin.h>
 #include <gtk/gtk_plugin.h>
-#include <local_notifier/local_notifier_plugin.h>
 #include <printing/printing_plugin.h>
-#include <realm/realm_plugin.h>
-#include <screen_retriever/screen_retriever_plugin.h>
+#include <screen_retriever_linux/screen_retriever_linux_plugin.h>
 #include <sentry_flutter/sentry_flutter_plugin.h>
 #include <smart_auth/smart_auth_plugin.h>
+#include <sqlite3_flutter_libs/sqlite3_flutter_libs_plugin.h>
 #include <tray_manager/tray_manager_plugin.h>
 #include <url_launcher_linux/url_launcher_plugin.h>
 #include <window_manager/window_manager_plugin.h>
@@ -26,9 +24,6 @@ void fl_register_plugins(FlPluginRegistry* registry) {
   g_autoptr(FlPluginRegistrar) amplify_db_common_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "AmplifyDbCommonPlugin");
   amplify_db_common_plugin_register_with_registrar(amplify_db_common_registrar);
-  g_autoptr(FlPluginRegistrar) cbl_flutter_ce_registrar =
-      fl_plugin_registry_get_registrar_for_plugin(registry, "CblFlutterCe");
-  cbl_flutter_ce_register_with_registrar(cbl_flutter_ce_registrar);
   g_autoptr(FlPluginRegistrar) desktop_webview_auth_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "DesktopWebviewAuthPlugin");
   desktop_webview_auth_plugin_register_with_registrar(desktop_webview_auth_registrar);
@@ -41,24 +36,21 @@ void fl_register_plugins(FlPluginRegistry* registry) {
   g_autoptr(FlPluginRegistrar) gtk_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "GtkPlugin");
   gtk_plugin_register_with_registrar(gtk_registrar);
-  g_autoptr(FlPluginRegistrar) local_notifier_registrar =
-      fl_plugin_registry_get_registrar_for_plugin(registry, "LocalNotifierPlugin");
-  local_notifier_plugin_register_with_registrar(local_notifier_registrar);
   g_autoptr(FlPluginRegistrar) printing_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "PrintingPlugin");
   printing_plugin_register_with_registrar(printing_registrar);
-  g_autoptr(FlPluginRegistrar) realm_registrar =
-      fl_plugin_registry_get_registrar_for_plugin(registry, "RealmPlugin");
-  realm_plugin_register_with_registrar(realm_registrar);
-  g_autoptr(FlPluginRegistrar) screen_retriever_registrar =
-      fl_plugin_registry_get_registrar_for_plugin(registry, "ScreenRetrieverPlugin");
-  screen_retriever_plugin_register_with_registrar(screen_retriever_registrar);
+  g_autoptr(FlPluginRegistrar) screen_retriever_linux_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "ScreenRetrieverLinuxPlugin");
+  screen_retriever_linux_plugin_register_with_registrar(screen_retriever_linux_registrar);
   g_autoptr(FlPluginRegistrar) sentry_flutter_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "SentryFlutterPlugin");
   sentry_flutter_plugin_register_with_registrar(sentry_flutter_registrar);
   g_autoptr(FlPluginRegistrar) smart_auth_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "SmartAuthPlugin");
   smart_auth_plugin_register_with_registrar(smart_auth_registrar);
+  g_autoptr(FlPluginRegistrar) sqlite3_flutter_libs_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "Sqlite3FlutterLibsPlugin");
+  sqlite3_flutter_libs_plugin_register_with_registrar(sqlite3_flutter_libs_registrar);
   g_autoptr(FlPluginRegistrar) tray_manager_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "TrayManagerPlugin");
   tray_manager_plugin_register_with_registrar(tray_manager_registrar);
