@@ -2,11 +2,16 @@
 import 'dart:io';
 
 import 'package:path/path.dart';
-import 'package:path_provider/path_provider.dart';
 import 'package:sqflite_common/sqflite.dart';
+import 'package:path_provider/path_provider.dart'
+    if (dart.library.io) 'package:path_provider/path_provider.dart'; // Conditional import
 
 mixin DatabasePath {
-  static Future<String>  getDatabaseDirectory() async {
+  static Future<String> getDatabaseDirectory() async {
+    if (isTestEnvironment()) {
+      return '.';
+    }
+
     if (Platform.isWindows) {
       final appDir = await getApplicationDocumentsDirectory();
       return join(appDir.path, '_db');
@@ -20,5 +25,9 @@ mixin DatabasePath {
       final appDir = await getApplicationDocumentsDirectory();
       return join(appDir.path, '_db');
     }
+  }
+
+  static bool isTestEnvironment() {
+    return const bool.fromEnvironment('FLUTTER_TEST_ENV') == true;
   }
 }
