@@ -1,4 +1,5 @@
 import 'dart:io';
+import 'package:brick_supabase/testing.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 import 'package:brick_offline_first_with_supabase/brick_offline_first_with_supabase.dart';
 import 'package:brick_sqlite/brick_sqlite.dart';
@@ -16,7 +17,6 @@ export 'package:brick_core/query.dart'
 
 const dbFileName = "flipper_v3.sqlite";
 const queueName = "brick_offline_queue.sqlite";
-
 
 class Repository extends OfflineFirstWithSupabaseRepository {
   static late Repository? _singleton;
@@ -54,11 +54,7 @@ class Repository extends OfflineFirstWithSupabaseRepository {
       databaseFactory:
           Platform.isWindows ? databaseFactoryFfi : databaseFactory,
       databasePath: queuePath,
-      onReattempt: (http.Request re, o) {
-        // print("Request status: $o");
-        print("here");
-        // print("Object: ${re.body}");
-      },
+      onReattempt: (http.Request re, o) {},
       onRequestException: (request, object) {
         // Deal with failed requests see https://github.com/GetDutchie/brick/issues/527
         print("Offline request exception: $request");
@@ -66,21 +62,33 @@ class Repository extends OfflineFirstWithSupabaseRepository {
       },
     );
 
-    final supabase = await Supabase.initialize(
-      url: supabaseUrl,
-      anonKey: supabaseAnonKey,
-      httpClient: client,
-    );
+    final SupabaseClient supabaseClient;
+    final mock = SupabaseMockServer(modelDictionary: supabaseModelDictionary);
+
+    if (DatabasePath.isTestEnvironment()) {
+      // Use the mocked client in a test environment
+      await mock.setUp();
+      supabaseClient =
+          SupabaseClient(mock.serverUrl, mock.apiKey, httpClient: client);
+    } else {
+      // Initialize the real Supabase client in a non-test environment
+      supabaseClient = (await Supabase.initialize(
+        url: supabaseUrl,
+        anonKey: supabaseAnonKey,
+        httpClient: client,
+      ))
+          .client;
+    }
 
     final provider = SupabaseProvider(
-      supabase.client,
+      supabaseClient,
       modelDictionary: supabaseModelDictionary,
     );
 
     _singleton = Repository._(
       supabaseProvider: provider,
       sqliteProvider: SqliteProvider(
-        dbPath,
+        DatabasePath.isTestEnvironment() ? inMemoryDatabasePath : dbPath,
         databaseFactory:
             Platform.isWindows ? databaseFactoryFfi : databaseFactory,
         modelDictionary: sqliteModelDictionary,
@@ -89,5 +97,39 @@ class Repository extends OfflineFirstWithSupabaseRepository {
       offlineRequestQueue: queue,
       memoryCacheProvider: MemoryCacheProvider(),
     );
+  }
+
+  Future<void> deleteUnprocessedRequests() async {
+    try {
+      // Retrieve unprocessed requests
+      final requests =
+          await offlineRequestQueue.requestManager.unprocessedRequests();
+
+      // Extract the primary key column name
+      final primaryKeyColumn =
+          offlineRequestQueue.requestManager.primaryKeyColumn;
+
+      // Create a list to hold the futures for deletion operations
+      final List<Future<void>> deletionFutures = [];
+
+      // Iterate through the unprocessed requests
+      for (final request in requests) {
+        // Retrieve the request ID using the primary key column
+        final requestId = request[primaryKeyColumn] as int;
+
+        // Add the deletion future to the list
+        deletionFutures.add(
+          offlineRequestQueue.requestManager
+              .deleteUnprocessedRequest(requestId),
+        );
+      }
+
+      // Wait for all deletion operations to complete
+      await Future.wait(deletionFutures);
+
+      print("All unprocessed requests have been deleted.");
+    } catch (e) {
+      print("An error occurred while deleting unprocessed requests: $e");
+    }
   }
 }
